RavenDB-24310 studio: rename serverwide -> serverWide for connection strings
Address review comment: the server-wide connection strings folder and
identifiers used lowercase 'serverwide', inconsistent with the sibling
serverWideAnalyzers/serverWideSorters. Renamed the directory, the
appUrl.forServerWideConnectionStrings helper and its route hash, the
manageServer menu entry (alias/require/route/dynamicHash), and all
appUrl call sites. Pre-existing backup-related identifiers left as-is.
