+++ src/Raven.Studio/typescript/common/appUrl.ts
@@ -259,8 +259,8 @@ class appUrl {
         return "#admin/settings/serverWideCustomSorters";
     }
 
-    static forServerwideConnectionStrings(): string {
-        return "#admin/settings/serverwideConnectionStrings";
+    static forServerWideConnectionStrings(): string {
+        return "#admin/settings/serverWideConnectionStrings";
     }
 
     static forDatabases(databasesUrlAction?: "compact" | "restore", databaseToCompact?: string, shardToCompact?: number): string {
