+++ src/Raven.Studio/typescript/components/pages/resources/manageServer/serverWideConnectionStrings/ServerWideConnectionStrings.stories.tsx
